csharp/IOBuffer.cs: Change IntPtr cast to long instead of int.
IntPtr is a platform specific type and is expected to have different
values on 32-bit or 64-bit systems. In C#, int is a 32-bit signed integer,
while a 64-bit integer is represented as long.

+++ bindings/csharp/IOBuffer.cs
@@ -134,10 +134,10 @@ namespace iio
         /// <remarks>The number of samples written will not exceed the size of the buffer.</remarks>
         public void fill(byte[] array)
         {
-            int length = (int) iio_buffer_end(buf) - (int) iio_buffer_start(buf);
+            long length = (long) iio_buffer_end(buf) - (long) iio_buffer_start(buf);
             if (length > array.Length)
                 length = array.Length;
-            Marshal.Copy(array, 0, iio_buffer_start(buf), length);
+            Marshal.Copy(array, 0, iio_buffer_start(buf), (int)length);
         }
     }
 }
